requirements.txt
Cython
numpy
torch
networkx
tqdm
scipy
geomstats
geoopt
biopython
python-Levenshtein
matplotlib